4-advanced-architectures/rag-complex/requirements.txt
streamlit>=1.11.0
pandas
chromadb
openai>=1.0.0
python-dotenv